Add `converse-rootview` plugin
Responsible for rendering container for all Converse UI.

By default it'll be inserted into the DOM, but the goal is to also have
it insertable at a later time.

This lets us decouple views and models and will enable us to only insert
the Converse UI on demand, while the headless part is already running
from the start.